Removed AbstractTxPool
